kernel のコンパイル警告修正 (gpt-5.6 luna)

+++ src/kernel/pci.cpp
@@ -273,10 +273,13 @@ namespace pci {
         const auto lo = static_cast<uint64_t>(bar);
         const auto hi = static_cast<uint64_t>(bar_upper);
         const auto ret = lo | (hi << 32);
-        Log(kDebug, "xHCI bus: 0x%llx, device: 0x%llx, function: 0x%llx\n", device.bus, device.device, device.function);
-        Log(kDebug, "xHCI BAR0 lo  : 0x%llx\n", lo);
-        Log(kDebug, "xHCI BAR0 hi  : 0x%llx\n", hi);
-        Log(kDebug, "xHCI BAR0 ret : 0x%llx\n", ret);
+        Log(kDebug, "xHCI bus: 0x%llx, device: 0x%llx, function: 0x%llx\n",
+            static_cast<unsigned long long>(device.bus),
+            static_cast<unsigned long long>(device.device),
+            static_cast<unsigned long long>(device.function));
+        Log(kDebug, "xHCI BAR0 lo  : 0x%llx\n", static_cast<unsigned long long>(lo));
+        Log(kDebug, "xHCI BAR0 hi  : 0x%llx\n", static_cast<unsigned long long>(hi));
+        Log(kDebug, "xHCI BAR0 ret : 0x%llx\n", static_cast<unsigned long long>(ret));
         return {
             ret,
             MAKE_ERROR(Error::kSuccess)
